requirements.txt
rich-argparse
requests
selenium
pillow
colorama
